Implements service for session creation and disabilitation
